Refactor opening of config files
When there are multiple default config files, open only the one with
highest priority -- this is the usual behaviour in applications,
otherwise it is impossible to unset a variable defined in the
least-priority config file. Combined with the write-out config file
feature, the user config file is expected to derive from the defaults
and set all required options.

Next, handle config files specified with the 'default' parameter of
ArgumentParser.add_argument() correctly. That is, check if the value
of the arg is the default value or user-specified value and require only
user-specified config file paths to exist. Note that multiple default
config files can be set with the default parameter, e.g.
    default=["foo.conf", "bar.conf"]
However the arg action is still required to be "store", so the user can
specify only one config file per arg.