Preserve priority of arguments when assembling the command line
The already_on_command_line() check does most of the decision making,
but this is not enough. For example if you have

    p.add_argument("--log-level", action="store",
            choices=LOG_LEVELS.keys(), default="info",
            help="the verbosity level for terminal logging")
    p.add_argument("-d", "--debug", action="store_const",
            const="debug", dest="log_level",
            help="shorthand for '--log-level debug'")
    p.add_argument("-q", "--quiet", action="store_const",
            const="warning", dest="log_level",
            help="shorthand for '--log-level warning'")

then all --log-level, --debug and --quiet can be passed multiple times,
the last actually taking effect (e.g. 'log-level = "info"' in the config
file and --debug on command line). This case is not detected by
already_on_command_line(). The commit does not break anything, but
brings more logic to the code.